+++ Sources/PaymentButtons/PayPalCreditButton.swift
@@ -20,19 +20,19 @@ public final class PayPalCreditButton: PaymentButton {
     /// Initialize a PayPalCreditButton
     /// - Parameters:
     ///   - insets: Edge insets of the button, defining the spacing of the button's edges relative to its content.
-    ///   - color: Color of the button. Default to dark blue if not provided.
-    ///   - edges: Edges of the button. Default to softEdges if not provided.
+    ///   - color: Color of the button. Default to gold if not provided.
+    ///   - shape: Shape of the button. Default to `.rounded` if not provided.
     ///   - size: Size of the button. Default to standard if not provided.
     public convenience init(
         insets: NSDirectionalEdgeInsets? = nil,
         color: Color = .darkBlue,
-        edges: PaymentButtonEdges = .softEdges,
+        shape: PaymentButtonShape = .rounded,
         size: PaymentButtonSize = .standard
     ) {
         self.init(
             fundingSource: PaymentButtonFundingSource.credit,
             color: color.color,
-            edges: edges,
+            shape: shape,
             size: size,
             insets: insets,
             label: nil
@@ -53,20 +53,20 @@ public extension PayPalCreditButton {
         /// Initialize a PayPalCreditButton
         /// - Parameters:
         ///   - insets: Edge insets of the button, defining the spacing of the button's edges relative to its content.
-        ///   - color: Color of the button. Default to dark blue if not provided.
-        ///   - edges: Edges of the button. Default to softEdges if not provided.
-        ///   - size: Size of the button. Default to standard if not provided.
+        ///   - color: Color of the button. Default to `.darkBlue` if not provided.
+        ///   - shape: Shape of the button. Default to `.rounded` if not provided.
+        ///   - size: Size of the button. Default to `.standard` if not provided.
         public init(
             insets: NSDirectionalEdgeInsets? = nil,
             color: PayPalCreditButton.Color = .darkBlue,
-            edges: PaymentButtonEdges = .softEdges,
+            shape: PaymentButtonShape = .rounded,
             size: PaymentButtonSize = .standard,
             _ action: @escaping () -> Void = { }
         ) {
             self.button = PayPalCreditButton(
                 fundingSource: PaymentButtonFundingSource.credit,
                 color: color.color,
-                edges: edges,
+                shape: shape,
                 size: size,
                 insets: insets,
                 label: nil